+++ www/client/src/Pages/Settings.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import LanguageSelector from "../components/LanguageSelector";
 import ColorSelector from "../components/ColorSelector";
+import CircleOfFifthsSelector from "../components/CircleOfFifthsSelector";
 import MidiOutPortSelector from "../components/MidiOutPortSelector";
 import MidiInPortSelector from '../components/MidiInPortSelector';
 import DynamicsController from "../components/DynamicsController";
 import TempoController from "../components/TempoController";
 import TranspositionController from "../components/TranspositionController";
+import KeyLabelToggle from "../components/KeyLabelToggle";
 
 const Settings: React.FC = () => {
 	return (
@@ -16,6 +18,8 @@ const Settings: React.FC = () => {
 		}}>
 			<ColorSelector />
 			<LanguageSelector />
+			<CircleOfFifthsSelector />
+			<KeyLabelToggle />
 			<MidiOutPortSelector />
 			<MidiInPortSelector />
 			<DynamicsController />